feat: refactor import paths for login and registration hooks
