+++ frontend/app/(auth)/_components/registration-form.tsx
@@ -23,7 +23,7 @@ import {
   registrationSchema,
 } from "@/schemas/auth/registration";
 
-import { useRegisterUser } from "@/app/(auth)/sign-up/actions/use-register-user";
+import { useRegisterUser } from "@/app/(auth)/sign-up/hooks/use-register-user";
 
 export default function RegistrationForm() {
   const [showPassword, setShowPassword] = useState(false);
